step1 et step2

+++ src/solver/Solver.cc
@@ -9,10 +9,10 @@ namespace solver{
 
 	}
 
-	float Solver::solveVerbose(){
+	double Solver::solveVerbose(){
 		std::cout << "Start solving problem... " << std::endl;
 		auto start = std::chrono::steady_clock::now();
-		float value = this->solveProblem();
+		double value = this->solveProblem();
 		auto end = std::chrono::steady_clock::now();
 		std::chrono::duration<double, std::milli> diff = end - start;
 
@@ -22,7 +22,7 @@ namespace solver{
 		return value;
 	}
 
-	float Solver::solve(){
+	double Solver::solve(){
 		return this->solveProblem();
 	}
 }
